+++ src/main/java/kimp/chat/repository/ChatRepository.java
@@ -1,6 +1,6 @@
 package kimp.chat.repository;
 
-import kimp.chat.dto.ChatLogDTO;
+import kimp.chat.dto.response.ChatLogResponseDto;
 import kimp.chat.entity.Chat;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -12,5 +12,5 @@ public interface ChatRepository extends MongoRepository<Chat, String> {
 
 
     @Query(value="{}", fields = "{chatID: 1, content: 1}")
-    List<ChatLogDTO> findAllByOrderByRegisted_atAsc(Pageable pageable);
+    List<ChatLogResponseDto> findAllByOrderByRegisted_atAsc(Pageable pageable);
 }
