+++ src/main/java/kimp/chat/service/ChatService.java
@@ -1,6 +1,6 @@
 package kimp.chat.service;
 
-import kimp.chat.dto.ChatLogDTO;
+import kimp.chat.dto.response.ChatLogResponseDto;
 import kimp.chat.entity.Chat;
 
 import java.util.List;
@@ -9,5 +9,5 @@ public interface ChatService {
 
     public Chat createChat(String chatID, String content);
 
-    public List<ChatLogDTO> getChatMessages(int page, int size);
+    public List<ChatLogResponseDto> getChatMessages(int page, int size);
 }
